reworked after review and fixed a failing test
